app-dd-ext/requirements.txt
Sanic==20.12.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-datadog
